refactor(web): 按业务域整理前端目录
