+++ docs/个人主页账户资料与自动打卡实施计划.md
@@ -81,9 +81,9 @@
 
 ### Task 8：前端
 
-- [ ] 修改 `frontend/web/src/apiClient.js`：增加 overview、昵称、头像和密码请求；运行 `npm run build`；单文件 commit。
-- [ ] 修改 `frontend/web/src/App.jsx`：增加 Profile 状态、动态日历、昵称/头像/密码交互、结果打卡提示和全局用户头像；运行 `npm run build`；单文件 commit。
-- [ ] 修改 `frontend/web/src/styles.css`：增加资料编辑、上传、密码弹窗和打卡提示样式；运行 `npm run build`；单文件 commit。
+- [ ] 修改 `frontend/web/src/infrastructure/http/apiClient.js`：增加 overview、昵称、头像和密码请求；运行 `npm run build`；单文件 commit。
+- [ ] 修改 `frontend/web/src/controller/App.jsx`：增加 Profile 状态、动态日历、昵称/头像/密码交互、结果打卡提示和全局用户头像；运行 `npm run build`；单文件 commit。
+- [ ] 修改 `frontend/web/src/common/styles.css`：增加资料编辑、上传、密码弹窗和打卡提示样式；运行 `npm run build`；单文件 commit。
 
 ### Task 9：接口与部署文档
