+++ docs/个人主页账户资料与自动打卡开发设计.md
@@ -106,7 +106,7 @@
    `idx_session_evaluation_scene_id (scene_id, created_at DESC)`。
 6. 旧对话清理只删除 `session_message` 和 `turn_evaluation`，不会删除
    `session_evaluation`，因此历史五维报告可以用于推导打卡。
-7. 前端个人主页、学习日历、统计卡和成就目前都在 `src/App.jsx`。
+7. 前端个人主页、学习日历、统计卡和成就目前都在 `frontend/web/src/controller/App.jsx`。
 8. 当前侧边栏和个人主页将 `teacher.image` 错当成用户头像。
 9. “修改密码”按钮目前没有事件处理。
 10. 当前项目不引入 Redis；头像对象存储通过 `ObjectStorageProvider` 抽象，由七牛云
@@ -635,21 +635,21 @@ boolean updatePasswordAndAuthVersion(
 
 | 动作 | 文件 | 当前定位/锚点 | 必须修改的内容 |
 | --- | --- | --- | --- |
-| 修改 | `frontend/web/src/apiClient.js` | 当前第 65 行 `getCurrentUser()` 和第 69 行偏好接口之间 | 增加 `getProfileOverview`、`updateProfile`、`uploadAvatar`、`changePassword` |
-| 修改 | `frontend/web/src/App.jsx` | 当前第 572 行 `AppShell` | 新增 `avatarUrl` 参数；第 585 行用用户头像或默认头像替代 `teacher.image` |
-| 修改 | `frontend/web/src/App.jsx` | 当前第 1194 行 `ResultModal` | `evaluation` 存在时展示“今日已自动打卡”状态 |
-| 修改 | `frontend/web/src/App.jsx` | 当前第 1968 行 `Profile` | 使用 Profile account；增加昵称编辑入口和头像上传入口 |
-| 修改 | `frontend/web/src/App.jsx` | `Profile` 和 `Overview` 组件 | 删除统计卡和节奏硬编码，读取 overview.statistics；成就常量保持不变 |
-| 修改 | `frontend/web/src/App.jsx` | 当前第 2006 行 `LearningCalendar` | 按月份调用后端并渲染 `checkedDates`；月份切换支持历史月份，不允许未来月份 |
-| 修改 | `frontend/web/src/App.jsx` | `Overview` | 周时长、训练记录数、连续天数和七日节奏全部渲染真实数据 |
-| 修改 | `frontend/web/src/App.jsx` | 当前第 2082 行 `Settings` 的“账户与隐私” | 给“修改密码”接入弹窗；“删除账户”保持不变 |
-| 修改 | `frontend/web/src/App.jsx` | 当前第 2096 行 `App` 状态与第 2175 行认证启动 | 增加 profile 状态；认证启动并行加载当前月份 overview；资料更新后同步状态 |
-| 修改 | `frontend/web/src/App.jsx` | 当前第 2287 行 `logout()` 附近 | 修改密码成功后复用清理 Token 和跳转登录逻辑 |
-| 修改 | `frontend/web/src/App.jsx` | 当前第 2373～2374 行 Profile/AppShell 渲染 | 向 Profile 和 AppShell 传 account、avatarUrl 及更新回调 |
-| 修改 | `frontend/web/src/styles.css` | 当前第 256～258 行 sidebar avatar | 保持头像裁切；增加加载失败/默认头像状态 |
-| 修改 | `frontend/web/src/styles.css` | `.profile-user`、`.profile-user__avatar`、`.profile-user__edit`、`.bars` | 放大圆形头像和账户文字；编辑按钮放到头像右上角；增加真实柱状图零值和日期样式 |
-| 修改 | `frontend/web/src/styles.css` | 当前第 810～838 行日历 | 保持现有视觉，只将 `is-practiced` 数据源改为报告日期 |
-| 修改 | `frontend/web/src/styles.css` | 当前第 879 行设置区域 | 增加修改密码弹窗状态样式 |
+| 修改 | `frontend/web/src/infrastructure/http/apiClient.js` | 当前第 65 行 `getCurrentUser()` 和第 69 行偏好接口之间 | 增加 `getProfileOverview`、`updateProfile`、`uploadAvatar`、`changePassword` |
+| 修改 | `frontend/web/src/controller/App.jsx` | 当前第 572 行 `AppShell` | 新增 `avatarUrl` 参数；第 585 行用用户头像或默认头像替代 `teacher.image` |
+| 修改 | `frontend/web/src/controller/App.jsx` | 当前第 1194 行 `ResultModal` | `evaluation` 存在时展示“今日已自动打卡”状态 |
+| 修改 | `frontend/web/src/controller/App.jsx` | 当前第 1968 行 `Profile` | 使用 Profile account；增加昵称编辑入口和头像上传入口 |
+| 修改 | `frontend/web/src/controller/App.jsx` | `Profile` 和 `Overview` 组件 | 删除统计卡和节奏硬编码，读取 overview.statistics；成就常量保持不变 |
+| 修改 | `frontend/web/src/controller/App.jsx` | 当前第 2006 行 `LearningCalendar` | 按月份调用后端并渲染 `checkedDates`；月份切换支持历史月份，不允许未来月份 |
+| 修改 | `frontend/web/src/controller/App.jsx` | `Overview` | 周时长、训练记录数、连续天数和七日节奏全部渲染真实数据 |
+| 修改 | `frontend/web/src/controller/App.jsx` | 当前第 2082 行 `Settings` 的“账户与隐私” | 给“修改密码”接入弹窗；“删除账户”保持不变 |
+| 修改 | `frontend/web/src/controller/App.jsx` | 当前第 2096 行 `App` 状态与第 2175 行认证启动 | 增加 profile 状态；认证启动并行加载当前月份 overview；资料更新后同步状态 |
+| 修改 | `frontend/web/src/controller/App.jsx` | 当前第 2287 行 `logout()` 附近 | 修改密码成功后复用清理 Token 和跳转登录逻辑 |
+| 修改 | `frontend/web/src/controller/App.jsx` | 当前第 2373～2374 行 Profile/AppShell 渲染 | 向 Profile 和 AppShell 传 account、avatarUrl 及更新回调 |
+| 修改 | `frontend/web/src/common/styles.css` | 当前第 256～258 行 sidebar avatar | 保持头像裁切；增加加载失败/默认头像状态 |
+| 修改 | `frontend/web/src/common/styles.css` | `.profile-user`、`.profile-user__avatar`、`.profile-user__edit`、`.bars` | 放大圆形头像和账户文字；编辑按钮放到头像右上角；增加真实柱状图零值和日期样式 |
+| 修改 | `frontend/web/src/common/styles.css` | 当前第 810～838 行日历 | 保持现有视觉，只将 `is-practiced` 数据源改为报告日期 |
+| 修改 | `frontend/web/src/common/styles.css` | 当前第 879 行设置区域 | 增加修改密码弹窗状态样式 |
 
 前端交互要求：
