+++ frontend/web/ROUTES.md
@@ -1,6 +1,6 @@
 # 前端分层路由约定
 
-路由定义与解析统一位于 `src/router.js`。页面组件只负责渲染和调用路径生成器，后端合并时可以按以下模块边界接入数据接口。
+路由定义与解析统一位于 `src/controller/router.js`。页面组件只负责渲染和调用路径生成器，后端合并时可以按以下模块边界接入数据接口。
 
 ## 公共与账户
