+++ frontend/web/index.html
@@ -8,6 +8,6 @@
   </head>
   <body class="notranslate" translate="no">
     <div id="root" class="notranslate" translate="no"></div>
-    <script type="module" src="/src/main.jsx"></script>
+    <script type="module" src="/src/controller/main.jsx"></script>
   </body>
 </html>
