+++ frontend/web/src/common/styles.css
