+++ frontend/web/src/component/common/NewtonsCradle.jsx
