+++ frontend/web/src/component/help/helpUtils.js
