+++ frontend/web/src/component/profile/AbilityTrendChart.jsx
