+++ frontend/web/src/component/profile/WeaknessRecommendations.jsx
