+++ frontend/web/src/domain/content/data.js
